+++ changelog.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.4.0...HEAD)
+# [Unreleased](https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.5.0...HEAD)
+
+# [v3.5.0](https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.4.0...v3.5.0)
+### Changed
+* Replaced sdk logger with genesys-cloud-client-logger
 ### Fixed
 * allow device switching in FireFox by stopping existing media before requesting new media